requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
# Add your database drivers here based on which databases you connect to:
pymysql==1.1.0        # For MySQL
# psycopg2-binary==2.9.9 # For PostgreSQL
# aiosqlite==0.19.0     # For async SQLite
